+++ library/core/src/main/java/com/google/android/exoplayer2/upstream/LoadErrorHandlingPolicy.java
@@ -15,28 +15,52 @@
  */
 package com.google.android.exoplayer2.upstream;
 
+import androidx.annotation.IntDef;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.source.LoadEventInfo;
 import com.google.android.exoplayer2.source.MediaLoadData;
 import com.google.android.exoplayer2.upstream.Loader.Callback;
 import com.google.android.exoplayer2.upstream.Loader.Loadable;
 import java.io.IOException;
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 
 /**
  * Defines how errors encountered by loaders are handled.
  *
  * <p>A loader that can choose between one of a number of resources can exclude a resource when a
- * load error occurs. In this case, {@link #getBlacklistDurationMsFor(int, long, IOException, int)}
- * defines whether the resource should be excluded. Exclusion will succeed unless all of the
- * alternatives are already excluded.
+ * load error occurs. In this case, {@link #getExclusionDurationMsFor(int, LoadErrorInfo)} defines
+ * whether the resource should be excluded for a given {@link FallbackType fallback type}, and if so
+ * for how long. If the policy indicates that a resource should be excluded, the loader will exclude
+ * it for the specified amount of time unless all of the alternatives for the given fallback type
+ * are already excluded.
  *
- * <p>When exclusion does not take place, {@link #getRetryDelayMsFor(int, long, IOException, int)}
- * defines whether the load is retried. An error that's not retried will always be propagated. An
- * error that is retried will be propagated according to {@link #getMinimumLoadableRetryCount(int)}.
+ * <p>When exclusion does not take place, {@link #getRetryDelayMsFor(LoadErrorInfo)} defines whether
+ * the load is retried. An error that's not retried will always be propagated. An error that is
+ * retried will be propagated according to {@link #getMinimumLoadableRetryCount(int)}.
  *
  * <p>Methods are invoked on the playback thread.
  */
 public interface LoadErrorHandlingPolicy {
+
+  /** Fallback type. One of {@link #FALLBACK_TYPE_LOCATION} or {@link #FALLBACK_TYPE_TRACK}. */
+  @Documented
+  @Retention(RetentionPolicy.SOURCE)
+  @IntDef({FALLBACK_TYPE_LOCATION, FALLBACK_TYPE_TRACK})
+  @interface FallbackType {}
+
+  /**
+   * Fallback type that is using exclusion of locations (i.e., multiple URLs through which the same
+   * data is accessible).
+   */
+  int FALLBACK_TYPE_LOCATION = 1;
+  /**
+   * Fallback type that is using exclusion of tracks (i.e., multiple URLs through which different
+   * representations of the same content are available; for example the same video encoded at
+   * different bitrates or resolutions).
+   */
+  int FALLBACK_TYPE_TRACK = 2;
 
   /** Holds information about a load task error. */
   final class LoadErrorInfo {
@@ -63,36 +87,17 @@ public interface LoadErrorHandlingPolicy {
     }
   }
 
-  /** @deprecated Implement {@link #getBlacklistDurationMsFor(LoadErrorInfo)} instead. */
-  @Deprecated
-  default long getBlacklistDurationMsFor(
-      int dataType, long loadDurationMs, IOException exception, int errorCount) {
-    throw new UnsupportedOperationException();
-  }
-
   /**
    * Returns the number of milliseconds for which a resource associated to a provided load error
-   * should be excluded, or {@link C#TIME_UNSET} if the resource should not be excluded.
+   * should be excluded for a given {@link FallbackType fallback type}, or {@link C#TIME_UNSET} if
+   * the resource should not be excluded.
    *
+   * @param fallbackType The {@link FallbackType fallback type} used for exclusion.
    * @param loadErrorInfo A {@link LoadErrorInfo} holding information about the load error.
    * @return The exclusion duration in milliseconds, or {@link C#TIME_UNSET} if the resource should
    *     not be excluded.
    */
-  @SuppressWarnings("deprecation")
-  default long getBlacklistDurationMsFor(LoadErrorInfo loadErrorInfo) {
-    return getBlacklistDurationMsFor(
-        loadErrorInfo.mediaLoadData.dataType,
-        loadErrorInfo.loadEventInfo.loadDurationMs,
-        loadErrorInfo.exception,
-        loadErrorInfo.errorCount);
-  }
-
-  /** @deprecated Implement {@link #getRetryDelayMsFor(LoadErrorInfo)} instead. */
-  @Deprecated
-  default long getRetryDelayMsFor(
-      int dataType, long loadDurationMs, IOException exception, int errorCount) {
-    throw new UnsupportedOperationException();
-  }
+  long getExclusionDurationMsFor(@FallbackType int fallbackType, LoadErrorInfo loadErrorInfo);
 
   /**
    * Returns the number of milliseconds to wait before attempting the load again, or {@link
@@ -106,14 +111,7 @@ public interface LoadErrorHandlingPolicy {
    * @return The number of milliseconds to wait before attempting the load again, or {@link
    *     C#TIME_UNSET} if the error is fatal and should not be retried.
    */
-  @SuppressWarnings("deprecation")
-  default long getRetryDelayMsFor(LoadErrorInfo loadErrorInfo) {
-    return getRetryDelayMsFor(
-        loadErrorInfo.mediaLoadData.dataType,
-        loadErrorInfo.loadEventInfo.loadDurationMs,
-        loadErrorInfo.exception,
-        loadErrorInfo.errorCount);
-  }
+  long getRetryDelayMsFor(LoadErrorInfo loadErrorInfo);
 
   /**
    * Called once {@code loadTaskId} will not be associated with any more load errors.